pyproject.toml
[tool.poetry]
name = "ardoredifferenziata"
version = "0.1.0"
description = ""
authors = ["Massimiliano Montaleone <massimilianomontaleone@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
schedule = "^1.1.0"
beautifulsoup4 = "^4.11.1"
pid = "^3.0.4"
python-telegram-bot = "^13.14"
holidays = "^0.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
